+++ src/components/MentalCalculation.svelte
@@ -1,13 +1,17 @@
 <script lang="ts">
-  import type { EquationType,CoundownTriggers } from "$lib/types";
+  import type { EquationType, CoundownTriggers } from "$lib/types";
   
   import CountdownTime from "./CountdownTime.svelte";
   import Equation from "./Equation.svelte";
   import CountdownBar from "./CountdownBar.svelte";
   import Keypad from "./Keypad.svelte";
 
-  let { equationArray }:{ equationArray:EquationType[] } = $props();
+  import Modal from "./Modal.svelte";
 
+  let { newGameTrigger=$bindable(), equationArray }:{newGameTrigger:boolean, equationArray:EquationType[] } = $props();
+
+  let globalTimer = 30;
+  
   let score = $state(0);
 
   let response:number = $state(0);
@@ -58,23 +62,34 @@
       wrongAnimating = false;
     }, 1000);
   }
+  function startNewGame(){
+    countdownTimeTriggers["start"]=true;
+    countdownBarTriggers["start"]=true;
+    score=0;
+    startNewEquation();
+  }
   // Lancement initial du jeu
   $effect(()=>{
+    if( newGameTrigger ){
+      newGameTrigger=false;
+      countdownTimeEnd=false;
+      startNewGame();
+    }
+  })
+  $effect(()=>{
     if(!countdownTimeEnd){
-      // Lancement des chronos
-      countdownTimeTriggers["start"]=true;
-      countdownBarTriggers["start"]=true;
-      score=0;
+      startNewGame();
     }
   });
   // 
   $effect(()=>{
-    
     if(!countdownTimeEnd && countdownBarEnd){
       // Affichage rouge
       wrongAnimation();
       // Lance une nouvelle équation
       startNewEquation();
+      // Réinitialisation de la réponse
+      response = 0;
     }else if( countdownTimeEnd ){
       countdownBarTriggers["pause"]=true;
     }
@@ -82,25 +97,25 @@
 </script>
 
 <div class="container">
-  <CountdownTime bind:triggers={countdownTimeTriggers} bind:end={countdownTimeEnd} duration={20}/>
+  <CountdownTime bind:triggers={countdownTimeTriggers} bind:end={countdownTimeEnd} duration={globalTimer}/>
   {#key winAnimating || wrongAnimating}
   <Equation {equation} {response} {winAnimating} {wrongAnimating}/>
   {/key}
   <CountdownBar bind:triggers={countdownBarTriggers} duration={3000} bind:end={countdownBarEnd}/>
   <Keypad bind:response={response} {verification}/>
   
-  {#if countdownTimeEnd}
-  <div class="startMentalCalculation">
-    <button onclick={()=>{ countdownTimeEnd = false}}>Start to calculate ?</button>
-    <div class="score">
-      {#if score > 0}
-        <div>Votre score est de {score} !</div>
-      {/if}
-    </div>
-  </div>
-  <!-- <div>FIN, Score : </div> -->
-  {/if}
+  {#if countdownTimeEnd} <div class="blur"> </div> {/if}
 </div>
+<Modal onclick={()=>{ countdownTimeEnd = false}} title={`Résolvez autant d'équation possible en ${globalTimer} s`} show={countdownTimeEnd}>
+  {#if score > 0}
+    <div class="score">Votre score est de {score} !</div>
+  {:else}
+    <div class="score">Obtenez votre score après votre tentative</div>
+  {/if}
+</Modal>
+
+
+
 
 <style>
   .container{
@@ -111,37 +126,17 @@
     background-color: #EFF3EA;
     border: solid 2px grey;
   }
-  .startMentalCalculation{
+  .blur{
     position: absolute;
     top: 0px;
     background-color: #bbb7b788;
     width:100%;
     height:100%;
     z-index: 10;
-
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-evenly;
   }
   .score{
     width: 80%;
-    height:100px;
-  }
-  .score div{
-    width: 100%;
-    height:100%;
-    background-color: #fdffec;
     text-align: center;
-    
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-
-  }
-  .startMentalCalculation button{
-    font-size: 28px;
-    padding: 20px 10px;
   }
   
 </style>